Migrated methods to redux actions.
