+++ src/constants/chat.js
@@ -1,0 +1,21 @@
+export const SET = "SET"
+export const UNSET = "UNSET"
+export const INCREMENT = "INCREMENT"
+export const DECREMENT = "DECREMENT"
+export const NOTIFY = "NOTIFY"
+export const NOTIFY_TIMED = "NOTIFY_TIMED"
+export const CLEAR_NOTIFY = "CLEAR_NOTIFY"
+export const PLAY_AUDIO = "PLAY_AUDIO"
+export const STOP_AUDIO = "STOP_AUDIO"
+export const RELOAD = "RELOAD"
+export const SET_CHOICE = "SET_CHOICE"
+export const TOGGLE_CHOICE = "TOGGLE_CHOICE"
+export const AUTHENTICATE = "AUTHENTICATE"
+export const LOGOUT = "LOGOUT"
+export const SERVICES_FIND = "SERVICES_FIND"
+export const SERVICES_GET = "SERVICES_GET"
+export const HANDLE_CHOICE_SELECTION = "HANDLE_CHOICE_SELECTION"
+export const LOAD_PATH = "LOAD_PATH"
+export const ADD_CHAT = "ADD_CHAT"
+export const TOGGLE_TYPING = "TOGGLE_TYPING"
+export const ADD_WITH_ANIM = "ADD_WITH_ANIM"
